requirements.txt
numpy
scipy>=1.5.0
networkx
pandas
matplotlib
seaborn
scikit-learn
pyyaml
